+++ client/src/App.tsx
@@ -6,8 +6,7 @@ import AuthenticateUsers from "./routes/AuthenticateUsers";
 import CollegeDatabase from "./routes/CollegeDatabase";
 import StudentDatabase from "./routes/StudentDatabase";
 import StudentProfileRoute from "./routes/StudentProfile";
-import PeerProfileViewRoute from "./routes/PeerProfileView";
-import MentorProfileViewRoute from "./routes/MentorProfile";
+// import MentorProfileViewRoute from "./routes/MentorProfile";
 import EditStudent from "./routes/EditStudent";
 import CollegeProfile from "./routes/CollegeProfile";
 import EditCollege from "./routes/EditCollege";
